backend/requirements.txt
Django
djangorestframework
python-dotenv
google-generativeai

django-cors-headers
